+++ auto-last-modified-time/install.sh
@@ -3,11 +3,12 @@
 # Copyright 2017 Fwolf <fwolf.aide+git-hooks@gmail.com>
 # Distributed under the MIT license.
 #
-# Last Modified: 2017-08-20T10:24:31+08:00, r28
+# Last Modified: 2017-08-22T22:55:55+08:00, r39
 #====================================================================
 
 
 SCRIPT_DIR=${0%/*}/
+PWD_BAK="$PWD"
 
 
 . "$SCRIPT_DIR"../inc/git-functions.sh
@@ -18,22 +19,14 @@ checkGitRoot
 findGitRepositoryDir "$GIT_ROOT"
 
 
-cp -r "$SCRIPT_DIR"../inc "$GIT_REPO_DIR"hooks/
-cp -r "$SCRIPT_DIR"../auto-last-modified-time "$GIT_REPO_DIR"hooks/
-rm "$GIT_REPO_DIR"hooks/auto-last-modified-time/install.sh
-cp "$SCRIPT_DIR"../fix-permissions.sh "$GIT_REPO_DIR"hooks/
-
-PWD_BAK="$PWD"
-cd "$GIT_REPO_DIR"hooks/
-chmod +x inc/*
-chmod +x auto-last-modified-time/*
-chmod +x fix-permissions.sh
+HOOK_NAME="auto-last-modified-time"
+copyHookScripts "$HOOK_NAME"
 
 
 # Script need explicit exit with a code, and check here for reject
-HOOK_CONTENT="\${0%/*}/auto-last-modified-time/auto-last-modified-time.sh
+HOOK_CONTENT="\${0%/*}/${HOOK_NAME}/${HOOK_NAME}.sh
 if [ 0 -ne \$? ]; then
-    echo Commit is rejected by auto-last-modified-time hook
+    echo Commit is rejected by ${HOOK_NAME} hook
     exit 1
 fi"