+++ phpcs/install.sh
@@ -3,11 +3,12 @@
 # Copyright 2017 Fwolf <fwolf.aide+git-hooks@gmail.com>
 # Distributed under the MIT license.
 #
-# Last Modified: 2017-08-20T10:52:12+08:00, r29
+# Last Modified: 2017-08-22T22:55:55+08:00, r39
 #====================================================================
 
 
 SCRIPT_DIR=${0%/*}/
+PWD_BAK="$PWD"
 
 
 . "$SCRIPT_DIR"../inc/git-functions.sh
@@ -18,22 +19,14 @@ checkGitRoot
 findGitRepositoryDir "$GIT_ROOT"
 
 
-cp -r "$SCRIPT_DIR"../inc "$GIT_REPO_DIR"hooks/
-cp -r "$SCRIPT_DIR"../phpcs "$GIT_REPO_DIR"hooks/
-rm "$GIT_REPO_DIR"hooks/phpcs/install.sh
-cp "$SCRIPT_DIR"../fix-permissions.sh "$GIT_REPO_DIR"hooks/
-
-PWD_BAK="$PWD"
-cd "$GIT_REPO_DIR"hooks/
-chmod +x inc/*
-chmod +x phpcs/*
-chmod +x fix-permissions.sh
+HOOK_NAME="phpcs"
+copyHookScripts "$HOOK_NAME"
 
 
 # Script need explicit exit with a code, and check here for reject
-HOOK_CONTENT="\${0%/*}/phpcs/phpcs.sh
+HOOK_CONTENT="\${0%/*}/${HOOK_NAME}/${HOOK_NAME}.sh
 if [ 0 -ne \$? ]; then
-    echo Commit is rejected by phpcs hook
+    echo Commit is rejected by ${HOOK_NAME} hook
     echo Notice: this hook only check staged status of file
     exit 1
 fi"