+++ set-objects-writable/install.sh
@@ -3,11 +3,12 @@
 # Copyright 2017 Fwolf <fwolf.aide+git-hooks@gmail.com>
 # Distributed under the MIT license.
 #
-# Last Modified: 2017-08-20T10:24:31+08:00, r28
+# Last Modified: 2017-08-22T22:55:55+08:00, r39
 #====================================================================
 
 
 SCRIPT_DIR=${0%/*}/
+PWD_BAK="$PWD"
 
 
 . "$SCRIPT_DIR"../inc/git-functions.sh
@@ -18,19 +19,11 @@ checkGitRoot
 findGitRepositoryDir "$GIT_ROOT"
 
 
-cp -r "$SCRIPT_DIR"../inc "$GIT_REPO_DIR"hooks/
-cp -r "$SCRIPT_DIR"../set-objects-writable "$GIT_REPO_DIR"hooks/
-rm "$GIT_REPO_DIR"hooks/set-objects-writable/install.sh
-cp "$SCRIPT_DIR"../fix-permissions.sh "$GIT_REPO_DIR"hooks/
-
-PWD_BAK="$PWD"
-cd "$GIT_REPO_DIR"hooks/
-chmod +x inc/*
-chmod +x set-objects-writable/*
-chmod +x fix-permissions.sh
+HOOK_NAME="set-objects-writable"
+copyHookScripts "$HOOK_NAME"
 
 
-HOOK_CONTENT="\${0%/*}/set-objects-writable/set-objects-writable.sh"
+HOOK_CONTENT="\${0%/*}/${HOOK_NAME}/${HOOK_NAME}.sh"
 
 installHook post-commit $HOOK_CONTENT 1l0ebpnpj11s5j5p
 installHook post-merge $HOOK_CONTENT 1l0ebpnpj11s5j5p